+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Version 0.7.0 - TBD
+### Added
+- `Service.emit(...)` can now also be called with custom events
 
 ## Version 0.6.3 - 2024-07-19
 ### Fixed