+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## Version 0.7.0 - TBD
 ### Added
 - `Service.emit(...)` can now also be called with custom events
+- `connect.to (ServiceClass)` as alternative to `'service'` string
+- `connect.to ('db')` returning `cds.DatabaseService`
 
 ## Version 0.6.3 - 2024-07-19
 ### Fixed