+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Version 0.7.0 - TBD
 
-## Version 0.6.4 - TBD
+## Version 0.6.4 - 2024-08-05
 ### Added
 - `Service.emit(...)` can now also be called with custom events
 - `Service.before(...)` and `Service.after(...)` now accept bound and unbound functions as parameter